+++ eslint.config.js
@@ -5,7 +5,6 @@ import react from 'eslint-plugin-react';
 import reactHooks from 'eslint-plugin-react-hooks';
 import reactRefresh from 'eslint-plugin-react-refresh';
 import prettier from 'eslint-config-prettier';
-import eslintPluginPrettier from 'eslint-plugin-prettier';
 import pluginQuery from '@tanstack/eslint-plugin-query';
 import globals from 'globals';
 
@@ -28,17 +27,16 @@ export default defineConfig([
       'react-hooks': reactHooks,
       'react-refresh': reactRefresh,
       '@typescript-eslint': tsPlugin,
-      prettier: eslintPluginPrettier,
       '@tanstack/query': pluginQuery,
     },
     rules: {
       ...react.configs.recommended.rules,
       ...reactHooks.configs['recommended-latest'].rules,
+      ...prettier.rules,
       'react/react-in-jsx-scope': 'off',
       'react/prop-types': 'off',
       'react-hooks/preserve-manual-memoization': 'off',
       '@typescript-eslint/no-unused-vars': ['error', { varsIgnorePattern: '^[A-Z_]' }],
-      'prettier/prettier': 'warn',
       '@tanstack/query/exhaustive-deps': 'error',
     },
     settings: {
@@ -71,19 +69,16 @@ export default defineConfig([
       'react-hooks': reactHooks,
       'react-refresh': reactRefresh,
       '@typescript-eslint': tsPlugin,
-      prettier: eslintPluginPrettier,
     },
     rules: {
+      ...prettier.rules,
       'react/react-in-jsx-scope': 'off',
       'react/prop-types': 'off',
       'react-hooks/preserve-manual-memoization': 'off',
       '@typescript-eslint/no-unused-vars': ['error', { varsIgnorePattern: '^[A-Z_]' }],
-      'prettier/prettier': 'warn',
     },
     settings: {
       react: { version: 'detect' },
     },
   },
-
-  prettier,
 ]);